+++ src/connections/sap.cpp
@@ -148,7 +148,7 @@ class SQLRSERVER_DLLSPEC sapcursor : public sqlrservercursor {
 		bool		outputBind(const char *variable, 
 						uint16_t variablesize,
 						char *value, 
-						uint16_t valuesize, 
+						uint32_t valuesize, 
 						int16_t *isnull);
 		bool		outputBind(const char *variable,
 						uint16_t variablesize,
@@ -223,7 +223,7 @@ class SQLRSERVER_DLLSPEC sapcursor : public sqlrservercursor {
 		uint16_t	paramindex;
 		CS_INT		*outbindtype;
 		char		**outbindstrings;
-		uint16_t	*outbindstringlengths;
+		uint32_t	*outbindstringlengths;
 		int64_t		**outbindints;
 		double		**outbinddoubles;
 		datebind	*outbinddates;
@@ -686,7 +686,7 @@ sapcursor::sapcursor(sqlrserverconnection *conn, uint16_t id) :
 	parameter=new CS_DATAFMT[maxbindcount];
 	outbindtype=new CS_INT[maxbindcount];
 	outbindstrings=new char *[maxbindcount];
-	outbindstringlengths=new uint16_t[maxbindcount];
+	outbindstringlengths=new uint32_t[maxbindcount];
 	outbindints=new int64_t *[maxbindcount];
 	outbinddoubles=new double *[maxbindcount];
 	outbinddates=new datebind[maxbindcount];
@@ -1096,7 +1096,7 @@ bool sapcursor::inputBind(const char *variable,
 bool sapcursor::outputBind(const char *variable, 
 				uint16_t variablesize,
 				char *value, 
-				uint16_t valuesize, 
+				uint32_t valuesize, 
 				int16_t *isnull) {
 
 	checkRePrepare();
